Refactor hasSize
